move command line argument parsing into separate file; move logging into config package
