+++ aries_cloudagent/config/default_logging_config.ini
