+++ aries_cloudagent/config/tests/test_logging.py
@@ -1,6 +1,6 @@
 from unittest import mock
 
-from aries_cloudagent import LoggingConfigurator
+from .. import logging as test_module
 
 
 class TestLoggingConfigurator:
@@ -8,19 +8,19 @@ class TestLoggingConfigurator:
     host_arg_value = "host"
     port_arg_value = "port"
 
-    @mock.patch("aries_cloudagent.logging.path.join", autospec=True)
-    @mock.patch("aries_cloudagent.logging.fileConfig", autospec=True)
+    @mock.patch.object(test_module.path, "join")
+    @mock.patch.object(test_module, "fileConfig", autospec=True)
     def test_configure_default(self, mock_file_config, mock_os_path_join):
-        LoggingConfigurator.configure()
+        test_module.LoggingConfigurator.configure()
 
         mock_file_config.assert_called_once_with(
             mock_os_path_join.return_value, disable_existing_loggers=False
         )
 
-    @mock.patch("aries_cloudagent.logging.path.join", autospec=True)
-    @mock.patch("aries_cloudagent.logging.fileConfig", autospec=True)
+    @mock.patch.object(test_module.path, "join")
+    @mock.patch.object(test_module, "fileConfig", autospec=True)
     def test_configure_path(self, mock_file_config, mock_os_path_join):
         path = "a path"
-        LoggingConfigurator.configure(path)
+        test_module.LoggingConfigurator.configure(path)
 
         mock_file_config.assert_called_once_with(path, disable_existing_loggers=False)
